tools, actions: Add `prerelease` branch (#27537)
The `prerelease` branch will be used during releases instead of the
release lock.

* Build workflow will run on pushes to "prerelease" in addition to
  "trunk".
  * This includes pushes to mirror repos.
* Autotagger will now autotag from "prerelease" in addition to "trunk".
* Slack ping will ping the releases channel for failures on the
  prerelease branch.
* Adjusted some text for `jetpack release` subcommands.
* `create-release-branch.sh` script will now expect to be run from the
  prerelease branch, not trunk.
* Pre-push hook will not prompt for change entries for the prerelease
  branch.
* Deletion of the prerelease branch in the monorepo will trigger
  deletion of the branch in mirrors too.
* Creation and deletion of the prerelease branch will also trigger
  Slack notifications.

Note this does not run tests on pushes to the prerelease branch. The
intent is that the branch will be merged into trunk in the end, which
will run the tests at that point.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3527197801