autotagger: Don't tag packages with alpha deps (#29908)
When cherry-picking packages to a release branch, releasing them, and
then backporting to trunk for possible auto-tagging, we need to avoid
the case where a package itself would be ok to tag but something it
depends on is not, as this results in the package having a dependency on
an -alpha version of the depended-on package.

This happened with #29891, for example: waf was auto-tagged at v0.11.1
with a dependency on connection 1.51.6-alpha, because in trunk
connection had some extra commits.

This also makes it somewhat easier to see if an auto-tagging actually
happened, by outputting a notice for GH's summary screen in all cases.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4632329175