Update to node 20 (#33818)
* Update to node 20

Calypso has updated to node 20, which means it's time for us to do so as
well.

Also, it has been two years since we switched from yarn to pnpm. Let's
clean up the dummy `.engines.yarn` and remove the doc it references.

Also, since pnpm is the only thing (as far as I know) we use that checks
package.json `.engines` and it's happy to check the monorepo root for
everything, let's remove the engines from all the projects'
package.jsons so we don't have to have such a giant PR in the future.
That means we can also get rid of the linting that syncs all the
projects' package.jsons and the special logic in the mirroring that
strips it out of the mirrors.

* assets: Update expected error message in test

* d'oh

* version bumps

---------

Co-authored-by: Igor Zinovyev <zinigor@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6866313594